Refactor: Split ContactForm into smaller components
Split the ContactForm component into smaller, more manageable components for improved maintainability and readability.
