+++ src/components/contact/ContactEmailField.tsx
@@ -1,0 +1,27 @@
+
+import { FormField, FormItem, FormLabel, FormControl, FormMessage } from "@/components/ui/form";
+import { Input } from "@/components/ui/input";
+import { UseFormReturn } from "react-hook-form";
+import { ContactFormValues } from "./form-schema";
+
+interface ContactEmailFieldProps {
+  form: UseFormReturn<ContactFormValues>;
+}
+
+export function ContactEmailField({ form }: ContactEmailFieldProps) {
+  return (
+    <FormField
+      control={form.control}
+      name="email"
+      render={({ field }) => (
+        <FormItem>
+          <FormLabel>Email Address</FormLabel>
+          <FormControl>
+            <Input placeholder="Enter your email" {...field} />
+          </FormControl>
+          <FormMessage />
+        </FormItem>
+      )}
+    />
+  );
+}
