+++ src/components/contact/ContactMessageField.tsx
@@ -1,0 +1,31 @@
+
+import { FormField, FormItem, FormLabel, FormControl, FormMessage } from "@/components/ui/form";
+import { Textarea } from "@/components/ui/textarea";
+import { UseFormReturn } from "react-hook-form";
+import { ContactFormValues } from "./form-schema";
+
+interface ContactMessageFieldProps {
+  form: UseFormReturn<ContactFormValues>;
+}
+
+export function ContactMessageField({ form }: ContactMessageFieldProps) {
+  return (
+    <FormField
+      control={form.control}
+      name="message"
+      render={({ field }) => (
+        <FormItem>
+          <FormLabel>Message</FormLabel>
+          <FormControl>
+            <Textarea 
+              placeholder="Describe your question or issue" 
+              className="min-h-[150px]" 
+              {...field} 
+            />
+          </FormControl>
+          <FormMessage />
+        </FormItem>
+      )}
+    />
+  );
+}
