+++ src/components/contact/ContactNameField.tsx
@@ -1,0 +1,27 @@
+
+import { FormField, FormItem, FormLabel, FormControl, FormMessage } from "@/components/ui/form";
+import { Input } from "@/components/ui/input";
+import { UseFormReturn } from "react-hook-form";
+import { ContactFormValues } from "./form-schema";
+
+interface ContactNameFieldProps {
+  form: UseFormReturn<ContactFormValues>;
+}
+
+export function ContactNameField({ form }: ContactNameFieldProps) {
+  return (
+    <FormField
+      control={form.control}
+      name="name"
+      render={({ field }) => (
+        <FormItem>
+          <FormLabel>Your Name</FormLabel>
+          <FormControl>
+            <Input placeholder="Enter your name" {...field} />
+          </FormControl>
+          <FormMessage />
+        </FormItem>
+      )}
+    />
+  );
+}
