+++ src/components/contact/ContactSubjectField.tsx
@@ -1,0 +1,38 @@
+
+import { FormField, FormItem, FormLabel, FormControl, FormMessage } from "@/components/ui/form";
+import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from "@/components/ui/select";
+import { UseFormReturn } from "react-hook-form";
+import { ContactFormValues, subjectOptions } from "./form-schema";
+
+interface ContactSubjectFieldProps {
+  form: UseFormReturn<ContactFormValues>;
+}
+
+export function ContactSubjectField({ form }: ContactSubjectFieldProps) {
+  return (
+    <FormField
+      control={form.control}
+      name="subject"
+      render={({ field }) => (
+        <FormItem>
+          <FormLabel>Subject</FormLabel>
+          <Select onValueChange={field.onChange} defaultValue={field.value}>
+            <FormControl>
+              <SelectTrigger>
+                <SelectValue placeholder="Select a subject" />
+              </SelectTrigger>
+            </FormControl>
+            <SelectContent>
+              {subjectOptions.map((option) => (
+                <SelectItem key={option.value} value={option.value}>
+                  {option.label}
+                </SelectItem>
+              ))}
+            </SelectContent>
+          </Select>
+          <FormMessage />
+        </FormItem>
+      )}
+    />
+  );
+}
